moved functions to utils/helper.js file
